backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
firebase-admin==6.4.0
boto3==1.34.40
httpx==0.27.0
python-dotenv==1.0.1
pydantic==2.6.1
google-cloud-speech==2.24.1 
